# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Audience:
    properties:
      building:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      department_uid:
        description: Может быть null
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      uuid:
        type: string
    type: object
  models.Discipline:
    properties:
      abbr:
        type: string
      actType:
        type: string
      fullName:
        type: string
      shortName:
        type: string
    type: object
  models.Group:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      department_uid:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      uuid:
        type: string
    type: object
  models.ScheduleItem:
    properties:
      audiences:
        items:
          $ref: '#/definitions/models.Audience'
        type: array
      created_at:
        type: string
      day:
        type: integer
      deleted_at:
        type: string
      discipline:
        $ref: '#/definitions/models.Discipline'
      endTime:
        type: string
      groups:
        items:
          $ref: '#/definitions/models.Group'
        type: array
      id:
        type: integer
      permission:
        type: string
      startTime:
        type: string
      stream:
        type: string
      teachers:
        items:
          $ref: '#/definitions/models.Teacher'
        type: array
      time:
        type: integer
      updated_at:
        type: string
      week:
        type: string
    type: object
  models.Teacher:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      middleName:
        type: string
      updated_at:
        type: string
      uuid:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для управления расписанием учебных занятий
  title: Автоматизированная система по ведению расписания учебных занятий
  version: "1.0"
paths:
  /get-data:
    get:
      consumes:
      - application/json
      description: Возвращает данные расписания из базы данных в формате JSON
      produces:
      - application/json
      responses:
        "200":
          description: Список элементов расписания
          schema:
            items:
              $ref: '#/definitions/models.ScheduleItem'
            type: array
        "500":
          description: 'error: Failed to fetch schedule items'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получение расписания
      tags:
      - GetData
  /get-group-schedule/{uuid}:
    get:
      consumes:
      - application/json
      description: Возвращает данные расписания конкретной группы из базы данных в
        формате JSON
      parameters:
      - description: UUID группы
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список элементов расписания
          schema:
            items:
              $ref: '#/definitions/models.ScheduleItem'
            type: array
        "500":
          description: 'error: Failed to fetch schedule items'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получение расписания группы
      tags:
      - GetData
  /hello:
    get:
      consumes:
      - application/json
      description: Проверяет, работает ли сервер и есть ли подключение к базе данных
      produces:
      - application/json
      responses:
        "200":
          description: Hello, World!
          schema:
            type: string
      summary: Проверка подключения
      tags:
      - Hello
  /insert-data:
    post:
      consumes:
      - application/json
      description: Вставляет данные расписания и экзаменов в базу данных
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Data inserted successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'errors: [error messages]'
          schema:
            additionalProperties: true
            type: object
      summary: Вставка данных
      tags:
      - InsertData
  /insert-group-schedule/{uuid}:
    post:
      consumes:
      - application/json
      description: Вставляет данные расписания и экзаменов для конкретной группы в
        базу данных
      parameters:
      - description: UUID группы
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Group schedule inserted successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'errors: [error messages]'
          schema:
            additionalProperties: true
            type: object
      summary: Вставка расписания группы
      tags:
      - InsertGroupSchedule
  /write-schedule:
    post:
      consumes:
      - application/json
      description: Сохраняет данные расписания в CSV файл
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Schedule written to file successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Failed to fetch schedule items" "error: Failed to write
            schedule to file'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Сохранение расписания
      tags:
      - WriteSchedule
swagger: "2.0"